refactor: address PR-4 reviewer-pattern checklist
- Drop residual ``starship_platform`` references in cutover_template.py
  docstring; document the actual contract (auth factory dispatches on
  Airflow's standard ``conn_type``: http for Astro/OSS, google_cloud_platform
  for GCC, aws for MWAA), matching common.py's Setup-endpoint mapping.
- Switch ``resolve_source_hook`` imports in operator + template from the
  bare ``...auth`` package to ``...auth.factory``; ``auth/__init__.py`` is
  intentionally empty in the split (PR-1 fixup), so the bare-package form
  would fail at runtime. Matches cutover.py's import style.
- Move ``resolve_source_hook`` import to module top-level in operators/
  starship.py and drop the speculative "deferred import" comment — cutover.py
  already imports at top-level, so the lazy form earned no keep.
- Add ``-> Any`` return type on ``StarshipCutoverMigrationOperator.execute``
  to match the file's convention.
